% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_rate.R
\name{inline_count_rate}
\alias{inline_count_rate}
\title{Make an inline count/rate calculation}
\usage{
inline_count_rate(x, count_fun = sum, rate_fun = mean, digits = 1)
}
\arguments{
\item{x}{A (typically logical) vector}

\item{count_fun}{A function to compute the count. Defaults to \code{sum}.}

\item{rate_fun}{A function to compute the rate. Defaults to \code{mean},}

\item{digits}{Round the rate digits}
}
\value{
A string with formatted count/rate
}
\description{
Make an inline count/rate calculation
}
\examples{
set.seed(123)
x <- rbinom(500, 1, 0.25)
inline_count_rate(x)
}
