% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable_aggregate_weighted_mean.R
\name{reactable_aggregate_weighted_mean}
\alias{reactable_aggregate_weighted_mean}
\title{Compute weighted averages in a reactable}
\usage{
reactable_aggregate_weighted_mean(x)
}
\arguments{
\item{x}{A object}
}
\value{
Another object
}
\description{
Compute weighted averages in a reactable
}
\examples{
reactable_aggregate_weighted_mean(5)
}
