% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectbl_agg_rm.R
\name{rectbl_agg_rm}
\alias{rectbl_agg_rm}
\title{Remove text from cells in a {reactable} aggregation}
\usage{
rectbl_agg_rm(col, threshold)
}
\arguments{
\item{col}{A string indicating the column to evaluate}

\item{threshold}{The threshold for \code{col} to remove text from the current cell}
}
\value{
A javascript function to be used in the \code{style} argument of \code{\link[reactable]{colDef}}
}
\description{
Removes the text from cells in which a specified column in that row exceeds a threshold
}
\examples{
if(require("reactable")) {

  # Make data set
  set.seed(123)
  n <- 10
  dat <-
    tibble::tibble(
      Group = rep(c("A", "B"), n / 2),
      Weight = rbinom(n, size = 500, prob = 1 / (1:n + 1)),
      `Mean of Groups` = rnorm(n, mean = 1:n),
      `Sum of Groups` = `Mean of Groups`
    )

  # Make table
  reactable(
    data = dat,
    groupBy = "Group",
    columns =
      list(
        Weight = colDef(style = rectbl_agg_rm("Weight", 400), aggregate = "sum"),
        `Mean of Groups` = colDef(style = rectbl_agg_rm("Weight", 400), aggregate = "mean"),
        `Sum of Groups` = colDef(aggregate = "sum")
      )
  )

}

}
