% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectbl_agg_wtd.R
\name{rectbl_agg_wtd}
\alias{rectbl_agg_wtd}
\title{Compute a weighted mean (or sum) in a {reactable} aggregation}
\usage{
rectbl_agg_wtd(wt_col, result = c("mean", "sum"))
}
\arguments{
\item{wt_col}{A string indicating the column to use for weighting}

\item{result}{Should the calculation return the weighted mean or sum?}
}
\value{
A JavaScript function to be used in the \code{aggregate} argument of \code{\link[reactable]{colDef}}
}
\description{
Compute a weighted mean (or sum) in a {reactable} aggregation
}
\examples{
if(require("reactable")) {

  # Make data set
  set.seed(123)
  n <- 10
  dat <-
    tibble::tibble(
      Group = rep(c("A", "B"), n / 2),
      Weight = rbinom(n, size = 500, prob = 1 / (1:n + 1)),
      `Mean of Groups` = rnorm(n, mean = 1:n),
      `Weighted Mean` = `Mean of Groups`,
      `Sum of Groups` = `Mean of Groups`,
      `Weighted Sum` = `Mean of Groups`
    )

  # Make table
  reactable(
    data = dat,
    groupBy = "Group",
    columns =
      list(
        Weight = colDef(aggregate = "sum"),
        `Mean of Groups` = colDef(aggregate = "mean"),
        `Weighted Mean` = colDef(aggregate = rectbl_agg_wtd("Weight")),
        `Sum of Groups` = colDef(aggregate = "sum"),
        `Weighted Sum` = colDef(aggregate = rectbl_agg_wtd("Weight", result = "sum"))
      )
  )

}

}
