% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rectbl_agg_wtd.R
\name{rectbl_agg_wtd}
\alias{rectbl_agg_wtd}
\title{Compute a weighted mean (or sum) in a {reactable} aggregation}
\usage{
rectbl_agg_wtd(wt_col, result = c("mean", "sum"))
}
\arguments{
\item{wt_col}{A string indicating the column to use for weighting}

\item{result}{Should the calculation return the weighted mean or sum?}
}
\value{
A javascript function to be used in the \code{agg} argument of \code{colDef}
}
\description{
Compute a weighted mean (or sum) in a {reactable} aggregation
}
\examples{
rectbl_agg_wtd("my_col")
}
